Reorganize main reconcile loop so it's not a single giant file
Split functions from the main devworkspace_controller.go file into
multiple files based on the role they play:
* lifecycle.go manages starting/stopping/failing workspaces
* setup.go manages setting up the controller
* status.go manages status updates
* util.go holds utility methods
