+++ controllers/workspace/util.go
@@ -1,0 +1,29 @@
+// Copyright (c) 2019-2021 Red Hat, Inc.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package controllers
+
+import (
+	"strings"
+
+	dw "github.com/devfile/api/v2/pkg/apis/workspaces/v1alpha2"
+	"github.com/google/uuid"
+)
+
+func getWorkspaceId(instance *dw.DevWorkspace) (string, error) {
+	uid, err := uuid.Parse(string(instance.UID))
+	if err != nil {
+		return "", err
+	}
+	return "workspace" + strings.Join(strings.Split(uid.String(), "-")[0:3], ""), nil
+}
